international_coal/requirements.txt
adbc-driver-manager==1.8.0
adbc-driver-sqlite==1.8.0
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
arro3-core==0.6.3
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.3
commonmark==0.9.1
connectorx==0.4.4
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
deltalake==1.1.4
Deprecated==1.2.18
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.1
faicons==0.2.2
fastexcel==0.16.0
fonttools==4.60.1
fsspec==2025.9.0
gevent==25.9.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.42.1
google-genai==1.47.0
great-tables==0.18.0
greenlet==3.2.4
gurobipy==12.0.3
h11==0.16.0
htmltools==0.6.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lightgbm==4.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
narwhals==2.6.0
nest-asyncio==1.6.0
numpy==2.3.3
openpyxl==3.1.5
optuna==4.5.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.1
polars==1.33.1
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
PuLP==3.3.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyiceberg==0.10.0
pyparsing==3.2.5
pyroaring==1.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.50.0
strictyaml==1.7.3
supertree==0.5.5
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.3
xlsx2csv==0.8.4
xlsxwriter==3.2.9
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
